Reorganize the code again to fix lint issues
