+++ src/app/components/Layout.tsx
@@ -5,7 +5,7 @@ import ErrorModal from '../errors/components/ErrorModal';
 import theme from '../theme';
 import AccessibilityButtonsWrapper from './AccessibilityButtonsWrapper';
 import NavBar from './NavBar';
-import { OnEsc } from './OnEsc';
+import OnEsc from './OnEsc';
 
 // tslint:disable-next-line:variable-name
 const Layout: SFC = ({ children }) => <AccessibilityButtonsWrapper>
