+++ src/app/components/OnEsc.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { dispatchKeyDownEvent, renderToDom } from '../../test/reactutils';
 import TestContainer from '../../test/TestContainer';
-import { OnEsc } from './OnEsc';
+import OnEsc from './OnEsc';
 
 describe('OnEsc', () => {
   it('doesn\'t throw when it runs without any registered callbacks', () => {
